use converter-object instead of convert-functions
